requirements.txt
langchain
langchain-community
google-generativeai
dotenv
faiss-cpu
sentence_transformers
streamlit